+++ pkg/samplerepo/assets/sampledata.go
@@ -1,0 +1,8 @@
+package assets
+
+import "embed"
+
+var (
+	//go:embed all:sample
+	SampleData embed.FS
+)