+++ clients/spark/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v0.7.3 - 2023-05-07
+
+Bug fixes:
+* Fix uncommitted gc to handle no uncommitted location (#5817)
+
+## v0.7.2 - 2023-05-02
+
+Bug fixes:
+* Fix uncommitted garbage collection bulk remove out of disk space (#5776)
+
 ## v0.7.1 - 2023-04-24
 
 === Performance improvements===